requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
tensorflow-hub==0.14.0
